Deduplicate database::FullOrder model::Order conversion
