+++ crates/shared/src/db_order_conversions.rs
@@ -1,5 +1,7 @@
 use {
     anyhow::{Context, Result},
+    bigdecimal::{BigDecimal, Zero},
+    chrono::Utc,
     database::{
         onchain_broadcasted_orders::OnchainOrderPlacementError as DbOnchainOrderPlacementError,
         orders::{
@@ -37,14 +39,14 @@ use {
     number_conversions::{big_decimal_to_big_uint, big_decimal_to_u256},
 };
 
-pub fn full_order_into_model_order(order: database::orders::FullOrder) -> Result<Order> {
-    let status = OrderStatus::Open;
+pub fn full_order_into_model_order(order: FullOrderDb) -> Result<Order> {
+    let status = calculate_status(&order);
     let pre_interactions = extract_interactions(&order, ExecutionTime::Pre)?;
     let post_interactions = extract_interactions(&order, ExecutionTime::Post)?;
     let ethflow_data = if let Some((refund_tx, user_valid_to)) = order.ethflow_data {
         Some(EthflowData {
             user_valid_to,
-            refund_tx_hash: refund_tx.map(|hash| H256::from(hash.0)),
+            refund_tx_hash: refund_tx.map(|hash| H256(hash.0)),
         })
     } else {
         None
@@ -261,5 +263,266 @@ pub fn signing_scheme_from(scheme: DbSigningScheme) -> SigningScheme {
         DbSigningScheme::EthSign => SigningScheme::EthSign,
         DbSigningScheme::Eip1271 => SigningScheme::Eip1271,
         DbSigningScheme::PreSign => SigningScheme::PreSign,
+    }
+}
+
+fn calculate_status(order: &FullOrderDb) -> OrderStatus {
+    match order.kind {
+        DbOrderKind::Buy => {
+            if is_buy_order_filled(&order.buy_amount, &order.sum_buy) {
+                return OrderStatus::Fulfilled;
+            }
+        }
+        DbOrderKind::Sell => {
+            if is_sell_order_filled(&order.sell_amount, &order.sum_sell, &order.sum_fee) {
+                return OrderStatus::Fulfilled;
+            }
+        }
+    }
+    if order.invalidated {
+        return OrderStatus::Cancelled;
+    }
+    if order.valid_to() < Utc::now().timestamp() {
+        return OrderStatus::Expired;
+    }
+    if order.presignature_pending {
+        return OrderStatus::PresignaturePending;
+    }
+    OrderStatus::Open
+}
+
+fn is_sell_order_filled(
+    amount: &BigDecimal,
+    executed_amount: &BigDecimal,
+    executed_fee: &BigDecimal,
+) -> bool {
+    if executed_amount.is_zero() {
+        return false;
+    }
+    let total_amount = executed_amount - executed_fee;
+    total_amount == *amount
+}
+
+fn is_buy_order_filled(amount: &BigDecimal, executed_amount: &BigDecimal) -> bool {
+    !executed_amount.is_zero() && *amount == *executed_amount
+}
+
+#[cfg(test)]
+mod tests {
+    use {super::*, chrono::Duration, database::byte_array::ByteArray};
+
+    #[test]
+    fn order_status() {
+        let valid_to_timestamp = Utc::now() + Duration::days(1);
+
+        let order_row = || FullOrderDb {
+            uid: ByteArray([0; 56]),
+            owner: ByteArray([0; 20]),
+            creation_timestamp: Utc::now(),
+            sell_token: ByteArray([1; 20]),
+            buy_token: ByteArray([2; 20]),
+            sell_amount: BigDecimal::from(1),
+            buy_amount: BigDecimal::from(1),
+            valid_to: valid_to_timestamp.timestamp(),
+            app_data: ByteArray([0; 32]),
+            fee_amount: BigDecimal::default(),
+            full_fee_amount: BigDecimal::default(),
+            kind: DbOrderKind::Sell,
+            class: DbOrderClass::Liquidity,
+            partially_fillable: true,
+            signature: vec![0; 65],
+            receiver: None,
+            sum_sell: BigDecimal::default(),
+            sum_buy: BigDecimal::default(),
+            sum_fee: BigDecimal::default(),
+            invalidated: false,
+            signing_scheme: DbSigningScheme::Eip712,
+            settlement_contract: ByteArray([0; 20]),
+            sell_token_balance: DbSellTokenSource::External,
+            buy_token_balance: DbBuyTokenDestination::Internal,
+            presignature_pending: false,
+            pre_interactions: Vec::new(),
+            post_interactions: Vec::new(),
+            ethflow_data: None,
+            onchain_user: None,
+            onchain_placement_error: None,
+            surplus_fee: Default::default(),
+            surplus_fee_timestamp: Default::default(),
+            executed_surplus_fee: Default::default(),
+            executed_solver_fee: Default::default(),
+        };
+
+        // Open - sell (filled - 0%)
+        assert_eq!(calculate_status(&order_row()), OrderStatus::Open);
+
+        // Open - sell (almost filled - 99.99%)
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Sell,
+                sell_amount: BigDecimal::from(10_000),
+                sum_sell: BigDecimal::from(9_999),
+                ..order_row()
+            }),
+            OrderStatus::Open
+        );
+
+        // Open - with presignature
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                signing_scheme: DbSigningScheme::PreSign,
+                presignature_pending: false,
+                ..order_row()
+            }),
+            OrderStatus::Open
+        );
+
+        // PresignaturePending - without presignature
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                signing_scheme: DbSigningScheme::PreSign,
+                presignature_pending: true,
+                ..order_row()
+            }),
+            OrderStatus::PresignaturePending
+        );
+
+        // Filled - sell (filled - 100%)
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Sell,
+                sell_amount: BigDecimal::from(2),
+                sum_sell: BigDecimal::from(3),
+                sum_fee: BigDecimal::from(1),
+                ..order_row()
+            }),
+            OrderStatus::Fulfilled
+        );
+
+        // Open - buy (filled - 0%)
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Buy,
+                buy_amount: BigDecimal::from(1),
+                sum_buy: BigDecimal::from(0),
+                ..order_row()
+            }),
+            OrderStatus::Open
+        );
+
+        // Open - buy (almost filled - 99.99%)
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Buy,
+                buy_amount: BigDecimal::from(10_000),
+                sum_buy: BigDecimal::from(9_999),
+                ..order_row()
+            }),
+            OrderStatus::Open
+        );
+
+        // Filled - buy (filled - 100%)
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Buy,
+                buy_amount: BigDecimal::from(1),
+                sum_buy: BigDecimal::from(1),
+                ..order_row()
+            }),
+            OrderStatus::Fulfilled
+        );
+
+        // Cancelled - no fills - sell
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                invalidated: true,
+                ..order_row()
+            }),
+            OrderStatus::Cancelled
+        );
+
+        // Cancelled - partial fill - sell
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Sell,
+                sell_amount: BigDecimal::from(2),
+                sum_sell: BigDecimal::from(1),
+                sum_fee: BigDecimal::default(),
+                invalidated: true,
+                ..order_row()
+            }),
+            OrderStatus::Cancelled
+        );
+
+        // Cancelled - partial fill - buy
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Buy,
+                buy_amount: BigDecimal::from(2),
+                sum_buy: BigDecimal::from(1),
+                invalidated: true,
+                ..order_row()
+            }),
+            OrderStatus::Cancelled
+        );
+
+        // Expired - no fills
+        let valid_to_yesterday = Utc::now() - Duration::days(1);
+
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                invalidated: false,
+                valid_to: valid_to_yesterday.timestamp(),
+                ..order_row()
+            }),
+            OrderStatus::Expired
+        );
+
+        // Expired - partial fill - sell
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Sell,
+                sell_amount: BigDecimal::from(2),
+                sum_sell: BigDecimal::from(1),
+                invalidated: false,
+                valid_to: valid_to_yesterday.timestamp(),
+                ..order_row()
+            }),
+            OrderStatus::Expired
+        );
+
+        // Expired - partial fill - buy
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                kind: DbOrderKind::Buy,
+                buy_amount: BigDecimal::from(2),
+                sum_buy: BigDecimal::from(1),
+                invalidated: false,
+                valid_to: valid_to_yesterday.timestamp(),
+                ..order_row()
+            }),
+            OrderStatus::Expired
+        );
+
+        // Expired - with pending presignature
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                signing_scheme: DbSigningScheme::PreSign,
+                invalidated: false,
+                valid_to: valid_to_yesterday.timestamp(),
+                presignature_pending: true,
+                ..order_row()
+            }),
+            OrderStatus::Expired
+        );
+
+        // Expired - for ethflow orders
+        assert_eq!(
+            calculate_status(&FullOrderDb {
+                invalidated: false,
+                ethflow_data: Some((None, valid_to_yesterday.timestamp())),
+                ..order_row()
+            }),
+            OrderStatus::Expired
+        );
     }
 }
